Add files via upload
This is one which I've done in a competition. The main motive of the competition is that we should increase our coding ability in various languages like HTML,CSS and Bootstrap etc... this is the code I've made and one more thing is that participants are instructed to do a wed as it is in the a screenshot that they send to us within 1:30 hours.